+++ classes/class-uagb-block-helper.php
@@ -211,11 +211,13 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 
 			// @codingStandardsIgnoreEnd
 
-			$desktop = UAGB_Helper::generate_css( $selectors, '#uagb-section-' . $id );
+			$base_selector = ( $attr['classMigrate'] ) ? '.uagb-block-' : '#uagb-section-';
 
-			$tablet = UAGB_Helper::generate_css( $t_selectors, '#uagb-section-' . $id );
+			$desktop = UAGB_Helper::generate_css( $selectors, $base_selector . $id );
 
-			$mobile = UAGB_Helper::generate_css( $m_selectors, '#uagb-section-' . $id );
+			$tablet = UAGB_Helper::generate_css( $t_selectors, $base_selector . $id );
+
+			$mobile = UAGB_Helper::generate_css( $m_selectors, $base_selector . $id );
 
 			$generated_css = array(
 				'desktop' => $desktop,
